Extract EventCategorySelect as functional component
